# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Drape API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /client/orders/{id}:
    get:
      operationId: getClientOrder
      tags: [client]
      summary: Get a single order with messages and review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientOrderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /producer/orders:
    get:
      operationId: getProducerOrders
      tags: [producer]
      summary: List all producer orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProducerOrderSummary"
  /producer/orders/{id}:
    get:
      operationId: getProducerOrder
      tags: [producer]
      summary: Get full order detail for producer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProducerOrderDetail"
    patch:
      operationId: updateProducerOrder
      tags: [producer]
      summary: Update order status, notes, or timeline
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProducerOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /producer/orders/{id}/messages:
    post:
      operationId: sendProducerMessage
      tags: [producer]
      summary: Send a message on an order as producer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required: [content]
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /producer/storefront:
    get:
      operationId: getProducerStorefront
      tags: [producer]
      summary: Get producer storefront/profile data
      responses:
        "200":
          description: Storefront data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProducerStorefront"
    patch:
      operationId: updateProducerStorefront
      tags: [producer]
      summary: Update producer storefront/profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProducerStorefrontBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    OrderStatus:
      type: string
      enum: [ENQUIRY, ACCEPTED, DEPOSIT_PAID, IN_PRODUCTION, FITTING, FINAL_PAYMENT, DELIVERED, COMPLETED, CANCELLED]
    TimelineEvent:
      type: object
      properties:
        date:
          type: string
        label:
          type: string
        completed:
          type: boolean
        note:
          type: ["string", "null"]
      required: [date, label, completed]
    ClientOrder:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        title:
          type: string
        description:
          type: ["string", "null"]
        agreedPrice:
          type: ["integer", "null"]
        currency:
          type: string
        depositPaid:
          type: boolean
        dueDate:
          type: ["string", "null"]
        estimatedDays:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        producerId:
          type: string
        producerName:
          type: ["string", "null"]
        producerStudioName:
          type: ["string", "null"]
        reviewed:
          type: boolean
      required: [id, status, title, currency, depositPaid, createdAt, updatedAt, producerId, reviewed]
    OrderMessage:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        senderId:
          type: string
        senderRole:
          type: string
        content:
          type: string
        readByClient:
          type: boolean
        readByProducer:
          type: boolean
        createdAt:
          type: string
      required: [id, orderId, senderId, senderRole, content, readByClient, readByProducer, createdAt]
    OrderReview:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        clientId:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, orderId, clientId, rating, createdAt]
    ClientOrderDetail:
      allOf:
        - $ref: "#/components/schemas/ClientOrder"
        - type: object
          properties:
            notes:
              type: ["string", "null"]
            timelineEvents:
              type: array
              items:
                $ref: "#/components/schemas/TimelineEvent"
            briefId:
              type: ["string", "null"]
            sessionId:
              type: ["string", "null"]
            producerInstagram:
              type: ["string", "null"]
            producerWhatsapp:
              type: ["string", "null"]
            messages:
              type: array
              items:
                $ref: "#/components/schemas/OrderMessage"
            review:
              oneOf:
                - $ref: "#/components/schemas/OrderReview"
                - type: "null"
...
    ProducerOrderSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
        agreedPrice:
          type: ["integer", "null"]
        currency:
          type: string
        depositPaid:
          type: boolean
        dueDate:
          type: ["string", "null"]
        estimatedDays:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        productionGuideAt:
          type: ["string", "null"]
        clientId:
          type: string
        clientName:
          type: ["string", "null"]
        clientEmail:
          type: string
      required: [id, title, status, currency, depositPaid, clientId, clientEmail, createdAt, updatedAt]
    ProducerOrderDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
        description:
          type: ["string", "null"]
        agreedPrice:
          type: ["integer", "null"]
        currency:
          type: string
        depositPaid:
          type: boolean
        dueDate:
          type: ["string", "null"]
        estimatedDays:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        timelineEvents:
          type: array
          items:
            type: object
        productionGuideAt:
          type: ["string", "null"]
        productionGuideContent:
          type: ["object", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        client:
          type: ["object", "null"]
        measurements:
          type: ["object", "null"]
        brief:
          type: ["object", "null"]
        lookbookImages:
          type: array
          items:
            type: object
        messages:
          type: array
          items:
            type: object
      required: [id, title, status, currency, depositPaid, createdAt, updatedAt, lookbookImages, messages]
    UpdateProducerOrderBody:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
        timelineEvents:
          type: array
          items:
            type: object
        agreedPrice:
          type: integer
        depositPaid:
          type: boolean
        estimatedDays:
          type: integer
        dueDate:
          type: string
...
    ProducerStorefront:
      type: object
      properties:
        profile:
          type: ["object", "null"]
        user:
          type: ["object", "null"]
      required: [profile, user]
    UpdateProducerStorefrontBody:
      type: object
      properties:
        studioName:
          type: string
        studioType:
          type: string
          enum: [SOLO, STUDIO]
        specialties:
          type: array
          items:
            type: string
        bio:
          type: string
        priceMin:
          type: integer
        priceMax:
          type: integer
        instagram:
          type: string
        portfolioUrls:
          type: array
          items:
            type: string
...